content_monitoring/app/urls.py
from django.urls import path,include
from rest_framework.routers import DefaultRouter
from .views import KeywordViewSet, FlagViewSet, scan
router=DefaultRouter()
router.register(r'keywords',KeywordViewSet)
router.register(r'flags',FlagViewSet)

urlpatterns=[
    path('',include(router.urls)),
    path('scan/', scan),]
