content_monitoring/requirements.txt
asgiref==3.11.1
blinker==1.9.0
click==8.3.0
colorama==0.4.6
contourpy==1.3.3
cycler==0.12.1
Django==6.0.3
djangorestframework==3.17.1
Flask==3.1.2
fonttools==4.62.1
itsdangerous==2.2.0
Jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.5.0
MarkupSafe==3.0.3
matplotlib==3.10.8
mlxtend==0.24.0
numpy==2.4.3
packaging==26.0
pandas==3.0.1
pillow==12.1.1
pyparsing==3.3.2
python-dateutil==2.9.0.post0
scikit-learn==1.8.0
scipy==1.16.3
six==1.17.0
sqlparse==0.5.5
threadpoolctl==3.6.0
tzdata==2025.3
Werkzeug==3.1.3
